feat(nexus info key): pass key to data plane
When creating a nexus for a volume, the control plane will pass the key
that should be used by the data plane to store the NexusInfo data
structure in the persistent store. In doing so, we reduce the chance
that the control plane and data plane keys do not match.

Note: When creating a nexus directly (i.e. a nexus not associated with a
volume) the control plane does not pass down the key.